requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
h11==0.14.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.5.1
uvicorn==0.31.0
